requirements.txt
Authomatic>=0.0.7
Flask>=0.8
flask_mongoengine
flask-debugtoolbar>=0.9.0
Flask-WTF
humanize>=0.5
itsdangerous>=0.23
mongoengine
newrelic>=2.16.0.12
premailer>=1.13
pycrypto>=2.6.1
pymongo>=2.6.3
pyquery>=1.2.8
python-binary-memcached>=0.22
pytz
requests>=1.0.4
#
#Flask-Funnel>=0.1.10